+++ Flagship/Utils/Utils.cs
@@ -1,5 +1,4 @@
-﻿using Flagship.Config;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -7,40 +6,12 @@ using System.Threading.Tasks;
 
 namespace Flagship.Utils
 {
-    internal static class Log
+    internal static class Utils
     {
-        public static void LogError(FlagshipConfig config, string message, string tag)
+        public static string TwoDigit(int value)
         {
-            try
-            {
-                if (config == null || config.LogManager == null ||  config.LogLevel< Enums.LogLevel.ERROR)
-                {
-                    return;
-                }
-
-                config.LogManager.Error(message, tag);
-            }
-            catch (Exception exception)
-            {
-                Console.WriteLine(exception.Message);
-            }
-        }
-
-        public static void LogInfo(FlagshipConfig config, string message, string tag)
-        {
-            try
-            {
-                if (config == null || config.LogManager==null || config.LogLevel < Enums.LogLevel.INFO)
-                {
-                    return;
-                }
-
-                config.LogManager.Info(message, tag);
-            }
-            catch (Exception exception)
-            {
-                Console.WriteLine(exception.Message);
-            }
+            var prefixValue = value < 10 ? "0" : null;
+            return $"{prefixValue}{value}";
         }
     }
 }
